+++ packages/web/src/javascripts/Components/NoteView/SuperEditor/SuperEditor.tsx
@@ -1,5 +1,12 @@
 import { WebApplication } from '@/Application/Application'
-import { ApplicationEvent, EditorLineHeight, isPayloadSourceRetrieved, PrefKey } from '@standardnotes/snjs'
+import {
+  ApplicationEvent,
+  classNames,
+  EditorFontSize,
+  EditorLineHeight,
+  isPayloadSourceRetrieved,
+  PrefKey,
+} from '@standardnotes/snjs'
 import { FunctionComponent, useCallback, useEffect, useRef, useState } from 'react'
 import { BlocksEditor, BlocksEditorComposer } from '@standardnotes/blocks-editor'
 import { ItemSelectionPlugin } from './Plugins/ItemSelectionPlugin/ItemSelectionPlugin'
@@ -29,6 +36,8 @@ import { SuperNoteMarkdownPreview } from './SuperNoteMarkdownPreview'
 import { ExportPlugin } from './Plugins/ExportPlugin/ExportPlugin'
 import GetMarkdownPlugin, { GetMarkdownPluginInterface } from './Plugins/GetMarkdownPlugin/GetMarkdownPlugin'
 import { AutoFocusPlugin } from '@lexical/react/LexicalAutoFocusPlugin'
+import { getPlaintextFontSize } from '@/Utils/getPlaintextFontSize'
+import ReadonlyPlugin from './Plugins/ReadonlyPlugin/ReadonlyPlugin'
 
 const NotePreviewCharLimit = 160
 
@@ -81,8 +90,6 @@ export const SuperEditor: FunctionComponent<Props> = ({
     })
   }, [application])
 
-  const [lineHeight, setLineHeight] = useState<EditorLineHeight>(PrefDefaults[PrefKey.EditorLineHeight])
-
   const handleChange = useCallback(
     async (value: string, preview: string) => {
       if (ignoreNextChange.current === true) {
@@ -129,10 +136,15 @@ export const SuperEditor: FunctionComponent<Props> = ({
     return disposer
   }, [controller, controller.item.uuid])
 
+  const [lineHeight, setLineHeight] = useState<EditorLineHeight>(PrefDefaults[PrefKey.EditorLineHeight])
+  const [fontSize, setFontSize] = useState<EditorFontSize | undefined>()
+
   const reloadPreferences = useCallback(() => {
     const lineHeight = application.getPreference(PrefKey.EditorLineHeight, PrefDefaults[PrefKey.EditorLineHeight])
+    const fontSize = application.getPreference(PrefKey.EditorFontSize, PrefDefaults[PrefKey.EditorFontSize])
 
     setLineHeight(lineHeight)
+    setFontSize(fontSize)
   }, [application])
 
   useEffect(() => {
@@ -144,7 +156,7 @@ export const SuperEditor: FunctionComponent<Props> = ({
   }, [reloadPreferences, application])
 
   return (
-    <div className="relative h-full w-full">
+    <div className="font-editor relative h-full w-full">
       <ErrorBoundary>
         <>
           <LinkingControllerProvider controller={linkingController}>
@@ -157,10 +169,13 @@ export const SuperEditor: FunctionComponent<Props> = ({
                 <BlocksEditor
                   onChange={handleChange}
                   ignoreFirstChange={true}
-                  className="relative h-full resize-none px-6 py-4 text-base focus:shadow-none focus:outline-none"
+                  className={classNames(
+                    'relative h-full resize-none px-4 py-4 focus:shadow-none focus:outline-none',
+                    lineHeight && `leading-${lineHeight.toLowerCase()}`,
+                    fontSize ? getPlaintextFontSize(fontSize) : 'text-base',
+                  )}
                   previewLength={NotePreviewCharLimit}
                   spellcheck={spellcheck}
-                  lineHeight={lineHeight}
                 >
                   <ItemSelectionPlugin currentNote={note.current} />
                   <FilePlugin />
@@ -176,6 +191,7 @@ export const SuperEditor: FunctionComponent<Props> = ({
                   <NodeObserverPlugin nodeType={BubbleNode} onRemove={handleBubbleRemove} />
                   <NodeObserverPlugin nodeType={FileNode} onRemove={handleBubbleRemove} />
                   <ExportPlugin />
+                  <ReadonlyPlugin note={note.current} />
                   {controller.isTemplateNote ? <AutoFocusPlugin /> : null}
                 </BlocksEditor>
               </BlocksEditorComposer>