+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.48.3](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.48.2...@standardnotes/mobile@3.48.3) (2022-11-30)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.48.2](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.48.1...@standardnotes/mobile@3.48.2) (2022-11-30)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.48.1](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.48.0...@standardnotes/mobile@3.48.1) (2022-11-29)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 # [3.48.0](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.47.14...@standardnotes/mobile@3.48.0) (2022-11-28)
 
 ### Features